Added comments to order db, fixed issue with charging commissions in chat, users/artist can now view orders

+++ database/migrations/2017_03_21_155042_create_commissions_table.php
@@ -18,7 +18,7 @@ class CreateCommissionsTable extends Migration
             $table->integer('user_id');
             $table->string('type');
             $table->text('description');
-            $table->float('price');
+            $table->decimal('price', 8, 2);
             $table->timestamps();
         });
     }
